--- app/src/main/java/org/sopt/sample/data/source/AuthDataSource.kt
@@ -1,11 +1,0 @@
-package org.sopt.sample.data.source
-
-import org.sopt.sample.data.model.SignInRequest
-import org.sopt.sample.data.model.SignInResponse
-import org.sopt.sample.data.model.SignUpRequest
-import org.sopt.sample.data.model.SignUpResponse
-
-interface AuthDataSource {
-    suspend fun signIn(signInRequest: SignInRequest): SignInResponse
-    suspend fun signUp(signUpRequest: SignUpRequest): SignUpResponse
-}
