--- app/src/main/java/org/sopt/sample/data/source/RegresDataSource.kt
@@ -1,7 +1,0 @@
-package org.sopt.sample.data.source
-
-import org.sopt.sample.data.model.ProfileListResponse
-
-interface RegresDataSource {
-    suspend fun getProfileList(): ProfileListResponse
-}
