--- app/src/main/java/org/sopt/sample/data/source/RegresDataSourceImpl.kt
@@ -1,9 +1,0 @@
-package org.sopt.sample.data.source
-
-import org.sopt.sample.data.api.RegresService
-import org.sopt.sample.data.model.ProfileListResponse
-
-class RegresDataSourceImpl(private val regresService: RegresService) : RegresDataSource {
-    override suspend fun getProfileList(): ProfileListResponse =
-        regresService.getProfileList()
-}
