+++ app/src/main/java/org/sopt/sample/data/source/remote/AuthDataSource.kt
@@ -1,4 +1,4 @@
-package org.sopt.sample.data.source
+package org.sopt.sample.data.source.remote
 
 import org.sopt.sample.data.api.AuthService
 import org.sopt.sample.data.model.SignInRequest
@@ -6,10 +6,10 @@ import org.sopt.sample.data.model.SignInResponse
 import org.sopt.sample.data.model.SignUpRequest
 import org.sopt.sample.data.model.SignUpResponse
 
-class AuthDataSourceImpl(private val authService: AuthService) : AuthDataSource {
-    override suspend fun signIn(signInRequest: SignInRequest): SignInResponse =
+class AuthDataSource(private val authService: AuthService) {
+    suspend fun signIn(signInRequest: SignInRequest): SignInResponse =
         authService.signIn(signInRequest)
 
-    override suspend fun signUp(signUpRequest: SignUpRequest): SignUpResponse =
+    suspend fun signUp(signUpRequest: SignUpRequest): SignUpResponse =
         authService.signUp(signUpRequest)
 }
